Move broken function tests to wip directory
